% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTransitionModels-package.R
\docType{package}
\name{BayesTransitionModels-package}
\alias{BayesTransitionModels}
\alias{BayesTransitionModels-package}
\title{BayesTransitionModels: Flexible spline-based estimation of transition models}
\description{
The fpemplus package implements Bayesian transition models that use flexible spline-based modeling techniques to estimate and project demographic and health indicators.
}
\author{
\strong{Maintainer}: Herbert Susmann \email{herbps10@gmail.com}

Authors:
\itemize{
  \item Leontine Alkema \email{lalkema@umass.edu}
  \item Boming Zhang \email{bomingzhang@umass.edu}
}

}
\keyword{internal}
