% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{fpem_coverage}
\alias{fpem_coverage}
\title{Calculate coverage of held-out observations.}
\usage{
fpem_coverage(fit, interval = 0.95, only_held_out = TRUE)
}
\arguments{
\item{fit}{fpemplus object}

\item{interval}{width of the predictive interval}

\item{only_held_out}{whether to only calculate coverage for held-out observations}
}
\description{
Calculate coverage of held-out observations.
}
