% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemplus-package.R
\docType{package}
\name{fpemplus-package}
\alias{fpemplus}
\alias{fpemplus-package}
\title{fpemplus: Flexible spline-based estimation of family planning indicators}
\description{
The fpemplus package implements a version of the Family Planning Estimation Model (FPEM) that uses flexible spline-based modeling techniques to estimate and project family planning indicators in populations.
}
\author{
\strong{Maintainer}: Herbert Susmann \email{herbps10@gmail.com}

Authors:
\itemize{
  \item Leontine Alkema \email{lalkema@umass.edu}
  \item Boming Zhang \email{bomingzhang@umass.edu}
}

}
\keyword{internal}
