% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemplus.R
\name{fpemplus}
\alias{fpemplus}
\title{Fit the fpemplus model to data.}
\usage{
fpemplus(
  data,
  y,
  se,
  year,
  source,
  area,
  start_year = NA,
  end_year = NA,
  t_star = NULL,
  model = "spline",
  num_knots = 7,
  spline_degree = 2,
  smoothing = TRUE,
  hierarchical_asymptote = c("intercept", area),
  hierarchical_splines = c("intercept", area),
  hierarchical_level = c("intercept", area),
  tau_prior = "normal(0, 1)",
  rho_prior = "uniform(0, 1)",
  extra_stan_data = list(),
  held_out = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{y}{column name of outcome.}

\item{se}{column name of outcome standard error.}

\item{year}{column name of outcome year.}

\item{source}{column name of data source.}

\item{area}{column name of the area of each observation}

\item{start_year}{start year of estimates.}

\item{end_year}{end year of estimates.}

\item{t_star}{reference year.}

\item{model}{which model to fit. Currently only "spline" is supported.}

\item{num_knots}{number of spline knots.}

\item{spline_degree}{spline degree. Degree 2 or 3 is supported.}

\item{smoothing}{whether to include smoothing component.}

\item{hierarchical_asymptote}{vector specifying hierarchical structure for asymptote (see Details).}

\item{hierarchical_splines}{vector specifying hierarchical structure for spline coefficients (see Details).}

\item{hierarchical_level}{vector specifying hierarchical structure for the level in reference year (see Details).}

\item{tau_prior}{Stan prior definition for the AR tau parameter}

\item{rho_prior}{Stan prior definition for the AR rho parameter}

\item{held_out}{binary vector indicating which observations are held out. Set to FALSE to hold out no observations.}

\item{...}{additional arguments for CmdStanModel::sample.}
}
\value{
fpemplus object.
}
\description{
Fit the fpemplus model to data.
}
\details{
Several area-specific parameters of the fpemplus model have hierarchical priors
assigned to them so that information can be shared between areas.
The package allows the structure of the hierarchical prior to be configured by the user
through the \code{hierarchical_asymptote}, \code{hierarchical_level}, and \code{hierarchical_splines} arguments.
These arguments expect a character vector that specifies a nesting hierarchical structure.
Each element of the vector must be either "intercept" or a column name in the dataset, where
"intercept" will add a global intercept for the parameter.
The vector must be in descending order in terms of the hierarchy: that is, it starts with
"intercept" and proceeds down the hierarchy.

For example, suppose we are fitting country-level data, where the dataset has columns
"name_country", "name_sub_region", and "name_region" containing the name of the country,
sub-region, and region that each observation belongs to. To specify that the spline coefficients
should be fitted with a hierarchical model in which countries are nested within sub-regions within regions within world,
we would use the argument
\code{hierarchical_splines = c("intercept", "name_region", "name_sub_region", "name_country")}.
}
