% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_data.R
\name{national_data}
\alias{national_data}
\title{National level dataset.}
\usage{
national_data(countries = c(), start_year = 1970)
}
\arguments{
\item{countries}{vector of countries}

\item{start_year}{first year}
}
\description{
National level dataset.
}
