% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_asymptote}
\alias{plot_asymptote}
\title{Plot the posterior of the asymptote parameter.}
\usage{
plot_asymptote(fit, hierarchical_level = NA, areas = c(), sort = FALSE)
}
\arguments{
\item{fit}{fpemplus object}

\item{hierarchical_level}{which hierarchical level to plot}

\item{areas}{vector areas to include in plot}

\item{sort}{boolean, whether to sort the areas by the posterior median asymptote}
}
\description{
Plots the posterior density of the asymptote parameter (\eqn{P}).
}
