% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_data_hyperparameters}
\alias{plot_data_hyperparameters}
\title{Plot data model parameters}
\usage{
plot_data_hyperparameters(fit, include_prior = TRUE)
}
\arguments{
\item{fit}{fpemplus object}

\item{include_prior}{whether to include prior distribution in plot}
}
\description{
Plots the posterior distribution of the non-sampling error parameters for each data source.
}
