% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hierarchical_sd}
\alias{plot_hierarchical_sd}
\title{Plot the standard deviation for each level of a hierarchical distribution}
\usage{
plot_hierarchical_sd(fit, variable)
}
\arguments{
\item{fit}{fpemplus object}

\item{variable}{hierarchical distribution to plot, one of "level", "asymptote", or "splines"}
}
\description{
Plot the standard deviation for each level of a hierarchical distribution
}
