% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_indicator}
\alias{plot_indicator}
\title{Plot posterior of the indicator.}
\usage{
plot_indicator(fit, areas = c(), color_sources = FALSE)
}
\arguments{
\item{fit}{fpemplus object}

\item{areas}{which areas to include in plot.}

\item{color_sources}{whether to color each data point by data source type}
}
\description{
Plots posterior quantiles of the latent indicator (\eqn{eta_{c,t}}).
}
