% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_level}
\alias{plot_level}
\title{Plot the level in the reference year parameter}
\usage{
plot_level(fit, hierarchical_level = NA, areas = c(), sort = FALSE)
}
\arguments{
\item{fit}{fpemplus object}

\item{hierarchical_level}{which hierarchical level to plot}

\item{areas}{areas to include in plot}

\item{sort}{boolean, whether to sort the areas by the posterior median level in reference year}
}
\description{
Plots the posterior density of the level in reference year parameter (\eqn{\Omega}).
}
