% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pit}
\alias{plot_pit}
\title{Plot Probability Integral Transform}
\usage{
plot_pit(fit, ...)
}
\arguments{
\item{fit}{fpemplus object}

\item{...}{additional arguments to geom_histogram}
}
\description{
Plot Probability Integral Transform
}
