% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot distribution of residuals}
\usage{
plot_residuals(fit, ...)
}
\arguments{
\item{fit}{fpemplus object}

\item{...}{additional arguments to geom_histogram}
}
\description{
Plot distribution of residuals
}
