% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_smoother}
\alias{plot_smoother}
\title{Plot posterior of the smoothing component.}
\usage{
plot_smoother(fit, areas = c())
}
\arguments{
\item{fit}{fpemplus object}

\item{areas}{which areas to include in plot.}
}
\description{
Plots posterior quantiles of the smoothing component (\eqn{epsilon_{c,t}}).
}
