% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_smoothing_hyperparameters}
\alias{plot_smoothing_hyperparameters}
\title{Plot smoothing model parameters}
\usage{
plot_smoothing_hyperparameters(fit, priors = NULL)
}
\arguments{
\item{fit}{fpemplus object}

\item{priors}{list with elements "tau" and "rho", each the density function of the corresponding prior distribution}
}
\description{
Plot smoothing model parameters
}
