% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_transition}
\alias{plot_transition}
\title{Plot transition functions for a level of hierarchy.}
\usage{
plot_transition(fit, hierarchical_level = NA, areas = c(), n = 0)
}
\arguments{
\item{fit}{fpemplus object}

\item{hierarchical_level}{which level of hierarchy to plot}

\item{areas}{which areas to include in the plot}

\item{n}{how many draws of the transition function to plot}
}
\description{
Plot transition functions for a level of hierarchy.
}
