% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpemplus.R
\name{print.fpemplus}
\alias{print.fpemplus}
\title{Print fpemplus object}
\usage{
\method{print}{fpemplus}(x, ...)
}
\arguments{
\item{x}{fpemplus object}

\item{...}{additional arguments (not currently used)}
}
\description{
Print fpemplus object
}
