% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.fpemplus}
\alias{print.summary.fpemplus}
\title{Print fpemplus summary}
\usage{
\method{print}{summary.fpemplus}(x, ...)
}
\arguments{
\item{x}{fpemplus summary object}

\item{...}{additional arguments (not currently used)}
}
\description{
Print fpemplus summary
}
