% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fpemplus}
\alias{summary.fpemplus}
\title{Summary of an fpemplus object}
\usage{
\method{summary}{fpemplus}(object, ...)
}
\arguments{
\item{object}{fpemplus object}

\item{...}{additional arguments (not currently used)}
}
\description{
Summary of an fpemplus object
}
