% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfrplus.R
\name{tfrplus}
\alias{tfrplus}
\title{Fit TFRplus model}
\usage{
tfrplus(
  data,
  y,
  year,
  area,
  start_year = NA,
  end_year = NA,
  t_star = NULL,
  model = "spline",
  num_knots = 7,
  spline_degree = 2,
  hierarchical_splines = c("intercept", area),
  tau_prior_sd = 1,
  rho_prior_sd = 1,
  held_out = FALSE,
  ...
)
}
\description{
Fit TFRplus model
}
